main.py
import os
from dotenv import load_dotenv
load_dotenv()

from langchain_ollama import ChatOllama
from langchain_core.tools import tool
from langchain_core.messages import HumanMessage

try:
    from langchain.agents import create_agent as _create_agent

    def build_agent(llm, tools):
        return _create_agent(llm, tools=tools)

except Exception:
    from langgraph.prebuilt import create_react_agent as _create_react_agent

    def build_agent(llm, tools):
        return _create_react_agent(llm, tools=tools)


# 1) Локальная модель через Ollama (модель и temperature можно задать через env)
_OLLAMA_MODEL = os.environ.get("OLLAMA_MODEL", "llama3.1")
_OLLAMA_TEMPERATURE = float(os.environ.get("OLLAMA_TEMPERATURE", "0"))
llm = ChatOllama(model=_OLLAMA_MODEL, temperature=_OLLAMA_TEMPERATURE)


# 2) Инструмент, который агент может вызывать
@tool
def multiply(a: int, b: int) -> int:
    """Умножает два целых числа a и b."""
    return a * b

# 3) Собираем агента
agent = build_agent(llm, tools=[multiply])

def ask_agent(text: str) -> str:
    """Отправляет одно сообщение агенту и возвращает финальный ответ."""
    result = agent.invoke({"messages": [HumanMessage(content=text)]})
    return result["messages"][-1].content

# 4) Интерактивный режим
print("Локальный агент запущен ✅")
print("Примеры: 'Сколько будет 7 умножить на 8?' или 'Умножь 12 на 5'")
print("Выход: exit / quit\n")

while True:
    user_text = input("Ты: ").strip()
    if user_text.lower() in ("exit", "quit"):
        print("Пока 👋")
        break

    try:
        answer = ask_agent(user_text)
        print("Агент:", answer, "\n")
    except Exception as e:
        print("Ошибка:", e)
        print("Проверь, что Ollama запущена и модель скачана: ollama list\n")
