requirements.txt
# Совместимость: Python 3.10–3.13 проверены; 3.14 — см. README.
# Версии подобраны для совместной работы без конфликтов.
langchain-ollama>=0.2.0,<0.3
langchain-core>=0.3.0,<0.4
langgraph>=0.2.0,<0.3
python-dotenv>=1.0.0
